3.jpa-hibernate/src/main/java/com/jpa/hibernate/jpahibernate/repository/CourseRepository.java
package com.jpa.hibernate.jpahibernate.repository;


import com.jpa.hibernate.jpahibernate.entity.Course;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

import javax.persistence.EntityManager;

@Repository
@Transactional
public class CourseRepository {

    @Autowired
    EntityManager em;

    public Course findById(Long id){
        return em.find(Course.class, id);
    }

    public void deleteById(Long id){
        Course course = findById(id);
        em.remove(course);
    }

    public Course save(Course course){
        if(course.getId() == null){
            //inser data
            em.persist(course);
        }else{
            //update data
            em.merge(course);
        }
        return course;
    }
}
